add method add(int, String val_code)

+++ src/main/java/maven/Rate.java
@@ -1,9 +1,11 @@
 package maven;
 
+import java.math.BigDecimal;
+
 public class Rate {
 //	public String no;
 //    public String effectiveDate;
-    public double mid;
+    public BigDecimal mid;
     public String currency;
     public String code;
